pyproject.toml
[project]
name = "allkinds-chat-bot"
version = "0.1.0"
description = "A service for anonymous communication between matched users"
requires-python = ">=3.9"

[tool.poetry]
name = "allkinds-chat-bot"
version = "0.1.0"
description = "A service for anonymous communication between matched users"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.3.0"
python-dotenv = "^1.0.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
sqlalchemy = "^2.0.27"
aiosqlite = "^0.20.0"
greenlet = "^3.0.0"
alembic = "^1.12.1"
pinecone-client = "^3.0.0"
openai = "^1.18.0"
requests = "^2.31.0"
psutil = "^5.9.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
chat-bot = "src.main:main_cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 